refactor: adapt to now-preffered file structures
