+++ api/bot.ts
@@ -1,0 +1,4 @@
+import { handleBotMessages } from '../src/bot'
+import { requestHandler } from '../src/index'
+
+export default requestHandler(handleBotMessages)
