+++ api/github.ts
@@ -1,0 +1,4 @@
+import { handleGitHubHook } from '../src/github'
+import { requestHandler } from '../src/index'
+
+export default requestHandler(handleGitHubHook)
