+++ api/interactive-components.ts
@@ -1,0 +1,4 @@
+import { handleInteractiveComponents } from '../src/bot'
+import { requestHandler } from '../src/index'
+
+export default requestHandler(handleInteractiveComponents)
