+++ api/oauth.ts
@@ -1,0 +1,4 @@
+import { handleOAuth } from '../src/bot'
+import { requestHandler } from '../src/index'
+
+export default requestHandler(handleOAuth)
