+++ api/paths.ts
@@ -1,0 +1,8 @@
+export const paths = {
+  GitHub: '/github',
+  bot: '/bot',
+  interactive: '/interactive-components',
+  oauth: '/oauth',
+  home: '/',
+  githubAssets: '/github-code-review-notifier', // asset path prefix on GitHub pages
+}
