+++ extra.d.ts
@@ -1,0 +1,5 @@
+import { NowRequest, NowResponse } from '@now/node'
+
+export type IncomingMessage = NowRequest
+
+export type RequestListener = (req: NowRequest, res: NowResponse) => void
