+++ global.d.ts
