--- src/routes/bot.ts
@@ -1,4 +1,0 @@
-import { handleBotMessages } from '../bot'
-import { requestHandler } from '../index'
-
-export default requestHandler(handleBotMessages)
