--- src/routes/github.ts
@@ -1,4 +1,0 @@
-import { handleGitHubHook } from '../github'
-import { requestHandler } from '../index'
-
-export default requestHandler(handleGitHubHook)
