--- src/routes/interactive.ts
@@ -1,4 +1,0 @@
-import { handleInteractiveComponents } from '../bot'
-import { requestHandler } from '../index'
-
-export default requestHandler(handleInteractiveComponents)
