--- src/routes/oauth.ts
@@ -1,4 +1,0 @@
-import { handleOAuth } from '../bot'
-import { requestHandler } from '../index'
-
-export default requestHandler(handleOAuth)
